+++ src/main/java/database/FileManager.java
@@ -1,4 +1,4 @@
-package filemanager;
+package database;
 
 import java.util.*;
 import java.lang.*;
