--- src/main/java/filemanager/UserInputReceiver.java
@@ -1,9 +1,0 @@
-package filemanager;
-
-public class UserInputReceiver {
-
-
-    private void print(Object o) {
-        System.out.println(o);
-    }
-}
