+++ src/main/java/mediator/Components/ListOfHabits.java
@@ -1,4 +1,6 @@
-package mediator;
+package mediator.Components;
+
+import mediator.Components.Component;
 
 public class ListOfHabits extends Component {
